chore: organize classes into packages
