+++ app/src/main/java/ru/otus/daggerhomework/App.kt
@@ -1,6 +1,8 @@
 package ru.otus.daggerhomework
 
 import android.app.Application
+import ru.otus.daggerhomework.di.ApplicationComponent
+import ru.otus.daggerhomework.di.DaggerApplicationComponent
 
 class App : Application() {
