+++ app/src/main/java/ru/otus/daggerhomework/di/ApplicationComponent.kt
@@ -1,4 +1,4 @@
-package ru.otus.daggerhomework
+package ru.otus.daggerhomework.di
 
 import android.content.Context
 import dagger.BindsInstance
