+++ app/src/main/java/ru/otus/daggerhomework/di/Qualifiers.kt
@@ -1,4 +1,4 @@
-package ru.otus.daggerhomework
+package ru.otus.daggerhomework.di
 
 import javax.inject.Qualifier
