+++ app/src/main/java/ru/otus/daggerhomework/di/Scopes.kt
@@ -1,4 +1,4 @@
-package ru.otus.daggerhomework
+package ru.otus.daggerhomework.di
 
 import javax.inject.Scope
