+++ app/src/main/java/ru/otus/daggerhomework/presentation/ColorGenerator.kt
@@ -1,4 +1,4 @@
-package ru.otus.daggerhomework
+package ru.otus.daggerhomework.presentation
 
 import android.graphics.Color
 import androidx.annotation.ColorInt
