+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/fragmentProducer"
-        android:name="ru.otus.daggerhomework.FragmentProducer"
+        android:name="ru.otus.daggerhomework.presentation.FragmentProducer"
         android:label="fragment_producer">
         <action
             android:id="@+id/action_fragmentProducer_to_fragmentReceiver"
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/fragmentReceiver"
-        android:name="ru.otus.daggerhomework.FragmentReceiver"
+        android:name="ru.otus.daggerhomework.presentation.FragmentReceiver"
         android:label="fragment_receiver">
         <action
             android:id="@+id/action_fragmentReceiver_to_fragmentProducer"
